requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
dateparser==1.1.1
fonttools==4.34.4
frozenlist==1.3.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.2
mplfinance==0.12.9b1
multidict==6.0.2
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.10
websockets==10.3
yarl==1.7.2
